Restructure the test runner and the server so that it is easier to split up files
