+++ server/plugins/api/settings.js
