+++ server/plugins/auth/register.js
