+++ server/plugins/auth/strategy.js
