+++ server/plugins/db/postgres.js
@@ -1,5 +1,5 @@
 const pg = require('pg');
-const config = require('../../config.js').pg;
+const config = require('../../../config.js').pg;
 
 exports.register = (server, options, next) => {
   server.app.pool = new pg.Pool(config);
