+++ server/plugins/db/redis.js
@@ -1,6 +1,6 @@
 const redis = require('redis');
 const bluebird = require('bluebird');
-const config = require('../../config.js').redis;
+const config = require('../../../config.js').redis;
 
 bluebird.promisifyAll(redis.RedisClient.prototype);
 bluebird.promisifyAll(redis.Multi.prototype);
