+++ test/server/api/index.test.js
@@ -1,0 +1,1 @@
+require('./collection.test.js');
