+++ test/server/auth/index.test.js
@@ -1,0 +1,5 @@
+require('./routes.test.js');
+require('./login.test.js');
+require('./register.test.js');
+require('./logout.test.js');
+require('./authflow.test.js');
