+++ test/server/auth/login.test.js
@@ -1,14 +1,14 @@
 const tape = require('tape');
 const assert = require('assert');
 
-const server = require('../../server/server.js');
+const server = require('../../../server/server.js');
 const redisCli = server.app.redisCli;
 const pool = server.app.pool;
 
-const flushDb = require('../helpers/flushDb.js')(pool, redisCli);
-const checkUserRegistered = require('../helpers/checkUserRegistered.js')(pool);
-const registerUser = require('../helpers/registerUser.js')(pool);
-const checkUserLoggedInWRedis = require('../helpers/checkUserLoggedInWRedis.js')(redisCli);
+const flushDb = require('../../helpers/flushDb.js')(pool, redisCli);
+const checkUserRegistered = require('../../helpers/checkUserRegistered.js')(pool);
+const registerUser = require('../../helpers/registerUser.js')(pool);
+const checkUserLoggedInWRedis = require('../../helpers/checkUserLoggedInWRedis.js')(redisCli);
 
 tape('POST :: /login', (t) => {
   const options = {
