+++ test/server/auth/register.test.js
@@ -1,12 +1,12 @@
 const tape = require('tape');
 const assert = require('assert');
 
-const server = require('../../server/server.js');
+const server = require('../../../server/server.js');
 const redisCli = server.app.redisCli;
 const pool = server.app.pool;
 
-const flushDb = require('../helpers/flushDb.js')(pool, redisCli);
-const checkUserRegistered = require('../helpers/checkUserRegistered.js')(pool);
+const flushDb = require('../../helpers/flushDb.js')(pool, redisCli);
+const checkUserRegistered = require('../../helpers/checkUserRegistered.js')(pool);
 
 tape('POST :: /register', (t) => {
   const options = {
