+++ test/test_helpers/index.test.js
@@ -1,0 +1,10 @@
+require('./checkUserRegistered.test.js');
+require('./checkUserLoggedInWRedis.test.js');
+require('./loginUserWRedis.test.js');
+require('./registerUser.test.js');
+require('./flushDb.test.js');
+require('./authenticate.test.js');
+require('./createCollection.test.js');
+require('./getCollections.test.js');
+require('./updateCollection.test.js'); // wip
+require('./getWords.test.js'); // wip
